#10 Web Services has been developed and some test codes has been fixed

+++ core/src/main/java/pl/com/revolut/impl/AccountServiceImpl.java
@@ -48,10 +48,10 @@ public class AccountServiceImpl extends StorageService<AccountId,Account> implem
     public void addOrUpdateAccount(Account account) throws NullParameterException {
         ServiceUtils.checkParameters(account);
         Account accountOld = super.getItem(account.getAccountId());
-        if(!accountOld.getCustomerId().equals(account.getCustomerId())) {
+        if(accountOld != null && !accountOld.getCustomerId().equals(account.getCustomerId())) {
             customerService.removeAccountFromCustomer(accountOld.getCustomerId(),accountOld.getAccountId());
-            customerService.addAccountToCustomer(account.getCustomerId(),account.getAccountId());
         }
+        customerService.addAccountToCustomer(account.getCustomerId(),account.getAccountId());
         super.addOrUpdateItem(account.getAccountId(),account);
     }
 
@@ -94,8 +94,11 @@ public class AccountServiceImpl extends StorageService<AccountId,Account> implem
     }
 
     @Override
-    public List<TransactionId> getTransactionsOfAccount(AccountId accountId) throws NullParameterException {
+    public List<TransactionId> getTransactionsOfAccount(AccountId accountId) throws NullParameterException, AccountException {
         ServiceUtils.checkParameters(accountId);
+        Account account = getAccountById(accountId);
+        if(accountId == null)
+            throw new AccountException();
         return accountIdToTransactionId.getOrDefault(accountId,new ArrayList<>());
     }
 }
